tests/requirements-test.txt
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-flask==1.2.0
coverage==7.3.0
Werkzeug==2.3.7
